Update FC3_HW2_Q1_Merkle-Hellman_Knapsack_Cryptosystem.py
Merkle-Hellman Knapsack Cryptosystem

Python code for key generation, encryption and decryption using the Merkle-Hellman Knapsack Cryptosystem. A very clear and well-written description of this algorithm can be found at the following link: 
http://en.wikipedia.org/wiki/Merkle–Hellman_knapsack_cryptosystem

An example run of my program appears below. It is assumed that the user will enter a string of fewer than 80 characters in length. If the user enters a longer string, inform them that the string entered is too long and ask the user to try again. 

Example execution:
Enter a string and I will encrypt it as single large integer.
Welcome to Data Structures and Algorithms
Clear text:
Welcome to Data Structures and Algorithms
Number of clear text bytes = 41
Welcome to Data Structures and Algorithms is encrypted as 317817076350145785260656997739621373931469119807217110529280649332942724728174120224095587842484635884305367159163265896397856056390477056525611299805189287161133817602808317806202994211855425964737022224210974561644533817599450919039345942975178915818105632933959978787221138943336909734004773052722627400695
Result of decryption: Welcome to Data Structures and Algorithms