FC3 HW2 Q2: Graph Coloring
A coloring of a graph is an assignment of a color to each vertex (node) of the graph so that no two vertices connected by an edge have the same color. The problem of coloring of graphs has been studied for many decades.  Unfortunately, coloring an arbitrary graph with as few colors as possible is one of a large class of problems called “NP-complete problems”, for which all known solutions are essentially of the type “try all possibilities”.  In the case of the coloring problem, “try all possibilities” means to try all assignments of colors to vertices at first one color, then two colors, then three, and so on.  It is generally believed that no algorithm to solve this problem can be substantially more efficient than this most obvious approach. While finding an optimal solution may require too much computer time, we may be able to get a ‘close to optimal’ solution in a reasonable amount of time. 

An application of graph coloring:

A data file contains a list of student class schedules.  The name of the data file should be hardwired to “StudentsCourses.txt”. The output of the program will be (a) a graph, created in NetworkX,  and (b) a final exam schedule. The format of the input file will always be as follows:
<student name>  <N>  <course name 1>  <course name 2> … <course name N> <return>
<student name>  <K>  <course name 1> <course name 2> … <course nameK> <return>
:
:
Each student’s name is in the form <last name>,<first name> with no intervening space characters . The course names are all 6 characters in length and are separated from each other by spaces.  Each course name refers to a single section of the course and for each course, only one section is offered. (It’s a small school.)

The number after the student’s name represents the number of courses the student is taking. There is no need to validate the file. We will assume that it has already been validated.

There is a maximum of 40 students on the file. No student may take more than 5 courses. There is a maximum of 20 different courses offered each term.

Your program must build a graph from the data on the file.  Process the graph with the greedy algorithm shown above and display a schedule of final exams.  Final exams must be scheduled in such a way that no student has a conflict.  For example, if Sue is taking ENG040 and MAT100 then the final exam for each of these courses must be scheduled at a different time.

The output of your program must be in the following form:
Final Exam Period 1
	<course name >
	<course name >
Final Exam Period 2
	<course name >
Final Exam Period P
             <course name >

ExampleData File: 
Smith, Al         2   mat100   che080
Jones,Sue         2   eng040   mat100
Bell,Amy         2   eng050   gym000
Gingrich,Fred  3   gym000  his098 eng040

The program would produce the following schedule:
Final Exam Period 1 =>  mat100  eng050  his098
Final Exam Period 2 => che080  eng040
Final Exam Period 3 => gym000

The program must also produce a colored graph created with NetworkX and displayed with matplotlib.